requirements.txt
pycrypto
hexdump